renamed package; added go mod
