+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/amazing0x41/jc-project
+
+go 1.13
+
+require github.com/stretchr/testify v1.4.0
